Moved xrt_deconvolve from response to util. Also moved test file, the PSF data files and the test data files to correct locations and edited test file to point to the correct locations.
